requirements.txt
flask
flask-cors
requests
google-genai
python-dotenv
gevent
gunicorn